Refaktorerer ut mapping av fellesDtoer til k9Format.
