+++ app/src/main/kotlin/no/nav/k9punsj/felles/ZoneUtils.kt
@@ -1,0 +1,7 @@
+package no.nav.k9punsj.felles
+
+import java.time.ZoneId
+
+object ZoneUtils {
+    val Oslo = ZoneId.of("Europe/Oslo")
+}
