+++ app/src/main/kotlin/no/nav/k9punsj/pleiepengerlivetssluttfase/PleiepengerLivetsSluttfaseSøknadDto.kt
@@ -34,78 +34,6 @@ data class PleiepengerLivetsSluttfaseSøknadDto(
     val trekkKravPerioder: Set<PeriodeDto> = emptySet(),
     val begrunnelseForInnsending: BegrunnelseForInnsending? = null) {
 
-    data class PleietrengendeDto(
-        val norskIdent: String? = null,
-        val foedselsdato: LocalDate? = null
-    )
-
-    data class ArbeidAktivitetDto(
-        val selvstendigNaeringsdrivende: SelvstendigNæringsdrivendeDto?,
-        val frilanser: FrilanserDto?,
-        val arbeidstaker: List<ArbeidstakerDto>?) {
-        data class SelvstendigNæringsdrivendeDto(
-            val organisasjonsnummer: String?,
-            val virksomhetNavn: String?,
-            val info: SelvstendigNæringsdrivendePeriodeInfoDto?) {
-            data class SelvstendigNæringsdrivendePeriodeInfoDto(
-                val periode: PeriodeDto?,
-                val virksomhetstyper: List<String>?,
-                val registrertIUtlandet: Boolean?,
-                val landkode: String?,
-                val regnskapsførerNavn: String?,
-                val regnskapsførerTlf: String?,
-                val bruttoInntekt: BigDecimal?,
-                val erNyoppstartet: Boolean?,
-                val erVarigEndring: Boolean?,
-                val endringInntekt: BigDecimal?,
-                @JsonFormat(pattern = "yyyy-MM-dd")
-                val endringDato: LocalDate?,
-                val endringBegrunnelse: String?
-            )
-        }
-
-        data class FrilanserDto(
-            @JsonFormat(pattern = "yyyy-MM-dd")
-            val startdato: String?,
-            @JsonFormat(pattern = "yyyy-MM-dd")
-            val sluttdato: String?,
-            val jobberFortsattSomFrilans: Boolean?
-        )
-
-        data class ArbeidstakerDto(
-            val norskIdent: String?,
-            val organisasjonsnummer: String?,
-            val arbeidstidInfo: ArbeidstidInfoDto?) {
-            data class ArbeidstidInfoDto(
-                val perioder: List<ArbeidstidPeriodeInfoDto>?) {
-                data class ArbeidstidPeriodeInfoDto(
-                    val periode: PeriodeDto?,
-                    val faktiskArbeidTimerPerDag: String?,
-                    val jobberNormaltTimerPerDag: String?,
-                    val faktiskArbeidPerDag: TimerOgMinutter? = faktiskArbeidTimerPerDag?.somDuration()?.somTimerOgMinutter()?.somTimerOgMinutterDto(),
-                    val jobberNormaltPerDag: TimerOgMinutter? = jobberNormaltTimerPerDag?.somDuration()?.somTimerOgMinutter()?.somTimerOgMinutterDto()
-                )
-            }
-        }
-    }
-
-    data class ArbeidstidDto(
-        val arbeidstakerList: List<ArbeidAktivitetDto.ArbeidstakerDto>?,
-        val frilanserArbeidstidInfo: ArbeidAktivitetDto.ArbeidstakerDto.ArbeidstidInfoDto?,
-        val selvstendigNæringsdrivendeArbeidstidInfo: ArbeidAktivitetDto.ArbeidstakerDto.ArbeidstidInfoDto?
-    )
-
-    data class BostederDto(
-        val periode: PeriodeDto?,
-        val land: String?
-    )
-
-    data class UtenlandsoppholdDto(
-        val periode: PeriodeDto?,
-        val land: String?,
-        val årsak: String?
-    )
-
     data class UttakDto(
         val periode: PeriodeDto?,
         val timerPleieAvPleietrengendePerDag: String?,
