+++ app/src/main/kotlin/no/nav/k9punsj/utils/StringUtils.kt
@@ -1,0 +1,9 @@
+package no.nav.k9punsj.utils
+
+object StringUtils {
+    fun String?.erSatt() = !isNullOrBlank()
+    fun String.blankAsNull() = when (isBlank()) {
+        true -> null
+        false -> this
+    }
+}
